+++ src/main/java/esn/entities/secondary/AbstractMessage.java
@@ -2,19 +2,15 @@ package esn.entities.secondary;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import esn.db.UserDAO;
 import esn.entities.User;
 import esn.entities.converters.JsonTimeSerializer;
-import org.hibernate.annotations.CreationTimestamp;
 
-import java.sql.Time;
-import java.sql.Timestamp;
 import java.util.Calendar;
 
 
 public abstract class AbstractMessage {
     @JsonIgnore
-    public int id;
+    public long id;
 
     @JsonIgnore
     public int userId;
@@ -33,7 +29,7 @@ public abstract class AbstractMessage {
 
 
 
-    public AbstractMessage(int id, String text, int orgId, User user) {
+    public AbstractMessage(long id, String text, int orgId, User user) {
         if (user == null) user = new User("Пользователь удалён", null);
         this.userId = user.getId();
         userName = user.getName();
@@ -67,7 +63,7 @@ public abstract class AbstractMessage {
         return imgUrl;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 }
