+++ src/main/java/esn/entities/secondary/Post.java
@@ -1,11 +1,10 @@
 package esn.entities.secondary;
 
-import esn.db.UserDAO;
 import esn.entities.User;
 
 public class Post extends AbstractMessage {
 
-    public Post(int id, String text, int orgId, User user) {
+    public Post(long id, String text, int orgId, User user) {
         super(id, text, orgId, user);
     }
 }
